refactor: move DatabaseProvider, DatabaseCollection, and LRUCacheTypes enums to a new types file
